requirements.txt
pycryptodome==3.15.0
pydrive==1.3.1
scapy==2.4.5